+++ Packages/com.github.kurotu.vrc-quest-tools/Editor/Models/ComponentRemover.cs
@@ -43,11 +43,7 @@ namespace KRT.VRCQuestTools.Models
         /// <returns>Unsupported components.</returns>
         internal virtual Component[] GetUnsupportedComponentsInChildren(GameObject gameObject, bool includeInactive)
         {
-            var components = gameObject.GetComponentsInChildren<Component>(includeInactive);
-            return components
-                .Where(c => c != null)
-                .Where(IsUnsupportedComponent)
-                .ToArray();
+            return VRCSDKUtility.UnsupportedComponentTypes.SelectMany(type => gameObject.GetComponentsInChildren(type, includeInactive)).ToArray();
         }
 
         /// <summary>