+++ Packages/com.github.kurotu.vrc-quest-tools/Editor/Utils/FinalIKUtility.cs
@@ -3,6 +3,9 @@
 // See LICENSE.txt file in the project root for full license information.
 // </copyright>
 
+using System;
+using System.Collections.Generic;
+using System.Linq;
 using UnityEngine;
 
 namespace KRT.VRCQuestTools.Utils
@@ -12,6 +15,13 @@ namespace KRT.VRCQuestTools.Utils
     /// </summary>
     internal static class FinalIKUtility
     {
+        /// <summary>
+        /// Gets FinalIK component types.
+        /// </summary>
+        internal static IEnumerable<Type> ComponentTypes => AppDomain.CurrentDomain.GetAssemblies()
+            .SelectMany(asm => asm.GetTypes())
+            .Where(IsFinalIKComponent);
+
         /// <summary>
         /// Whether a component is from FinalIK.
         /// </summary>